refactor: moves back many files from shared to core.
