+++ src/app/core/abstractions/basic-rundown.service.ts
@@ -1,5 +1,5 @@
 import { Observable } from 'rxjs'
-import { BasicRundown } from '../../core/models/basic-rundown'
+import { BasicRundown } from '../models/basic-rundown'
 
 export abstract class BasicRundownService {
   public abstract fetchBasicRundowns(): Observable<BasicRundown[]>
