+++ src/app/core/core.module.ts
@@ -13,10 +13,16 @@ import { RundownEntityService } from './services/models/rundown-entity.service'
 import { SegmentEntityService } from './services/models/segment-entity.service'
 import { PartEntityService } from './services/models/part-entity.service'
 import { BasicRundownEntityService } from './services/models/basic-rundown-entity.service'
-import { Logger } from './abstractions/logger.service'
-import { Tv2LoggerService } from './services/tv2-logger.service'
 import { RundownTimingService } from './services/rundown-timing.service'
 import { RundownTimingContextStateService } from './services/rundown-timing-context-state.service'
+import { ShowStyleVariantService } from './abstractions/show-style-variant.service'
+import { HttpShowStyleVariantService } from './services/http/http-show-style-variant.service'
+import { RundownService } from './abstractions/rundown.service'
+import { HttpRundownService } from './services/http/http-rundown.service'
+import { BasicRundownService } from './abstractions/basic-rundown.service'
+import { HttpBasicRundownService } from './services/http/http-basic-rundown.service'
+import { EntityParser } from './abstractions/entity-parser.service'
+import { ZodEntityParser } from './parsers/zod-entity-parser.service'
 
 @NgModule({
   declarations: [],
@@ -28,7 +34,10 @@ import { RundownTimingContextStateService } from './services/rundown-timing-cont
     BasicRundownStateService,
     ShowStyleVariantStateService,
     { provide: RundownEventParser, useClass: ZodRundownEventParser },
-    { provide: Logger, useClass: Tv2LoggerService },
+    { provide: ShowStyleVariantService, useClass: HttpShowStyleVariantService },
+    { provide: RundownService, useClass: HttpRundownService },
+    { provide: BasicRundownService, useClass: HttpBasicRundownService },
+    { provide: EntityParser, useClass: ZodEntityParser },
     RundownEntityService,
     BasicRundownEntityService,
     SegmentEntityService,
