+++ src/app/core/services/basic-rundown-state.service.ts
@@ -4,7 +4,7 @@ import { RundownEventObserver } from './rundown-event-observer.service'
 import { EventSubscription } from '../../event-system/abstractions/event-observer.service'
 import { ConnectionStatusObserver } from './connection-status-observer.service'
 import { BasicRundown } from '../models/basic-rundown'
-import { BasicRundownService } from '../../shared/abstractions/basic-rundown.service'
+import { BasicRundownService } from '../abstractions/basic-rundown.service'
 import { RundownCreatedEvent, RundownEvent, RundownUpdatedEvent } from '../models/rundown-event'
 import { BasicRundownEntityService } from './models/basic-rundown-entity.service'
 import { Logger } from '../abstractions/logger.service'
