+++ src/app/core/services/http/http-basic-rundown.service.ts
@@ -1,12 +1,12 @@
 import { Injectable } from '@angular/core'
 import { HttpClient } from '@angular/common/http'
-import { HttpErrorService } from './http-error.service'
+import { HttpErrorService } from '../../../shared/services/http/http-error.service'
 import { catchError, map, Observable } from 'rxjs'
 import { BasicRundownService } from '../../abstractions/basic-rundown.service'
-import { BasicRundown } from '../../../core/models/basic-rundown'
+import { BasicRundown } from '../../models/basic-rundown'
 import { EntityParser } from '../../abstractions/entity-parser.service'
 import { environment } from '../../../../environments/environment'
-import { HttpResponse } from './http-response'
+import { HttpResponse } from '../../../shared/services/http/http-response'
 
 @Injectable()
 export class HttpBasicRundownService implements BasicRundownService {
