+++ src/app/core/services/http/test/http-basic-rundown.service.spec.ts
@@ -1,6 +1,6 @@
 import { HttpBasicRundownService } from '../http-basic-rundown.service'
 import { HttpClient } from '@angular/common/http'
-import { HttpErrorService } from '../http-error.service'
+import { HttpErrorService } from '../../../../shared/services/http/http-error.service'
 import { instance, mock } from '@typestrong/ts-mockito'
 import { EntityParser } from '../../../abstractions/entity-parser.service'
