+++ src/app/core/services/http/test/http-show-style-variant.service.spec.ts
@@ -1,14 +1,14 @@
-import { HttpRundownService } from '../http-rundown.service'
-import { HttpErrorService } from '../http-error.service'
+import { HttpErrorService } from '../../../../shared/services/http/http-error.service'
 import { instance, mock } from '@typestrong/ts-mockito'
 import { HttpClient } from '@angular/common/http'
 import { EntityParser } from '../../../abstractions/entity-parser.service'
+import { HttpShowStyleVariantService } from '../http-show-style-variant.service'
 
-describe('HttpRundownService', () => {
+describe('HttpConfigurationService', () => {
   it('should be created', () => {
     const mockedHttpClient = mock<HttpClient>()
     const mockedHttpErrorService = mock<HttpErrorService>()
     const mockedEntityParser = mock<EntityParser>()
-    expect(new HttpRundownService(instance(mockedHttpClient), instance(mockedHttpErrorService), instance(mockedEntityParser))).toBeTruthy()
+    expect(new HttpShowStyleVariantService(instance(mockedHttpClient), instance(mockedHttpErrorService), instance(mockedEntityParser))).toBeTruthy()
   })
 })
