+++ src/app/core/services/show-style-variant-state.service.spec.ts
@@ -1,5 +1,5 @@
 import { instance, mock } from '@typestrong/ts-mockito'
-import { ShowStyleVariantService } from '../../shared/abstractions/show-style-variant.service'
+import { ShowStyleVariantService } from '../abstractions/show-style-variant.service'
 import { ShowStyleVariantStateService } from './show-style-variant-state.service'
 
 describe('ShowStyleVariantStateService', () => {
