+++ src/app/rundown-overview/components/rundown-overview/rundown-overview.component.ts
@@ -6,7 +6,7 @@ import { DialogService } from '../../../shared/services/dialog.service'
 import { Color } from '../../../shared/enums/color'
 import { BasicRundownStateService } from '../../../core/services/basic-rundown-state.service'
 import { SubscriptionLike } from 'rxjs'
-import { RundownService } from '../../../shared/abstractions/rundown.service'
+import { RundownService } from '../../../core/abstractions/rundown.service'
 import { IconButton, IconButtonSize } from '../../../shared/enums/icon-button'
 import { Logger } from '../../../core/abstractions/logger.service'
 import { RundownTimingType } from '../../../core/enums/rundown-timing-type'
