+++ src/app/rundown-view/components/rundown/rundown.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing'
 import { RundownComponent } from './rundown.component'
 import { instance, mock } from '@typestrong/ts-mockito'
 import { RundownStateService } from '../../../core/services/rundown-state.service'
-import { RundownService } from '../../../shared/abstractions/rundown.service'
+import { RundownService } from '../../../core/abstractions/rundown.service'
 import { DialogService } from '../../../shared/services/dialog.service'
 import { RundownTimingContextStateService } from '../../../core/services/rundown-timing-context-state.service'
 import { PartEntityService } from '../../../core/services/models/part-entity.service'
