+++ src/app/rundown-view/components/segment-context-menu/segment-context-menu.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core'
 import { ContextMenuOption } from '../../../shared/abstractions/context-menu-option'
 import { Part } from '../../../core/models/part'
-import { RundownService } from '../../../shared/abstractions/rundown.service'
+import { RundownService } from '../../../core/abstractions/rundown.service'
 import { Segment } from '../../../core/models/segment'
 
 @Component({
